finsyncR: An R Package to synchronize fish and invertebrate datasets from federal sources in the United States

finsyncR (fish and invertebrate synchronizer in R) flexibly acquires, processes, and integrates national-level aquatic biomonitoring datasets in the US, with a focus on fish and macroinvertebrates sampled in rivers and streams. The package streamlines the process of retrieving and working with these data, making the process of utilizing these data easier for researchers. The data resources used in this package are from United States Environmental Protection Agency’s National Aquatic Resources Surveys, namely the National River and Streams Assessment, and United States Geological Survey’s BioData.

Data from USGS BioData were obtained from Pete Ruhl (pmruhl@usgs.gov), and data from the National Aquatic Resource Surveys are directly downloaded by the package from the NARS website.

Users should cite the aforementioned datasets as follows:

U.S. Geological Survey, 2020. BioData - Aquatic bioassessment data for the Nation: U.S. Geological Survey database, accessed 17 December 2020, at https://doi.org/10.5066/F77W698B

U.S. Environmental Protection Agency. 2006. National Aquatic Resource Surveys. Wadeable Streams Assessment 2004 (data and metadata files). Available from U.S. EPA web page: https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.

U.S. Environmental Protection Agency. 2016. National Aquatic Resource Surveys. National Rivers and Streams Assessment 2008-2009 (data and metadata files). Available from U.S. EPA web page: https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.

U.S. Environmental Protection Agency. 2020. National Aquatic Resource Surveys. National Rivers and Streams Assessment 2013-2014 (data and metadata files). Available from U.S. EPA web page: https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.

U.S. Environmental Protection Agency. 2022. National Aquatic Resource Surveys. National Rivers and Streams Assessment 2018-2019 (data and metadata files). Available from U.S. EPA web page: https://www.epa.gov/national-aquatic-resource-surveys/data-national-aquatic-resource-surveys.

Metadata can be found in the "Metadata" vignette: vignette("Metadata", package = "finsyncR").

A compendium on the correct use of the finsyncR package and its output datasets can be found in the "Getting Started" vignette: vignette("GettingStarted", package = "finsyncR").

Of note: The approach provided in this package concerning the way the USEPA National Rivers and Streams Assessment data are treated is not necessarily the same as the approach used in the National Rivers and Streams Assessment national reports.

Disclaimer: The United States Environmental Protection Agency (EPA) GitHub project code is provided on an "as is" basis and the user assumes responsibility for its use. EPA has relinquished control of the information and no longer has responsibility to protect the integrity , confidentiality, or availability of the information. Any reference to specific commercial products, processes, or services by service mark, trademark, manufacturer, or otherwise, does not constitute or imply their endorsement, recommendation or favoring by EPA. The EPA seal and logo shall not be used in any manner to imply endorsement of any commercial product or activity by EPA or the United States Government.

See also

Author

Maintainer: Michael Mahon mahon.michael@epa.gov

Authors:

Samantha Rumschlag rumschlag.samantha@epa.gov

Devin Jones devin.k.jones@gmail.com

Ryan Hill hill.ryan@epa.gov

Terry Brown brown.terryn@epa.gov [reviewer]

Ethan Brown ebrown23@nd.edu

Stefan Kunz stkunz@uni-landau.de